fix more tests, pointing at type registry instead of catalog
note items added via OSGi will no longer show up in calls to search the catalog;
callers should use type registry instead.

at this point most things are done except:

* REST calls to catalog won't show things added via bundles (as they are in TypeRegistry)
* TEMPLATE items aren't supported
* LOCATION items aren't put/removed from BasicLocationRegistry (but I think that is no longer needed?)

need to fix those then can declare victory on this!
